+++ src/core/structs/hash_table/vectors/statistics/statistics_functions.rs
@@ -11,7 +11,7 @@ use crate::core::structs::hash_table::vectors::hash_vec::{
 /// * `index` - index of bucket
 pub(in crate::core::structs::hash_table) fn statistics_add_actions<K, V, M>(
     hash_vec: &mut M,
-    index: u64,
+    index: usize,
 ) where
     V: Default + Eq,
     K: Default + Eq,
@@ -21,9 +21,9 @@ pub(in crate::core::structs::hash_table) fn statistics_add_actions<K, V, M>(
     if let Some(len) = bucket_len {
         if len > hash_vec.get_max_len() {
             hash_vec.get_statistics_mut().update(len);
-            hash_vec.get_statistics_mut().add_bucket(index as usize);
+            hash_vec.get_statistics_mut().add_bucket(index);
         } else if len == hash_vec.get_statistics().max_length {
-            hash_vec.get_statistics_mut().add_bucket(index as usize);
+            hash_vec.get_statistics_mut().add_bucket(index);
         }
     }
     hash_vec.get_statistics_mut().size += 1;
@@ -38,23 +38,21 @@ pub(in crate::core::structs::hash_table) fn statistics_add_actions<K, V, M>(
 /// * `index` - index of bucket
 pub(in crate::core::structs::hash_table) fn statistics_remove_actions<K, V, M>(
     hash_vec: &mut M,
-    index: u64,
+    index: usize,
 ) where
     V: Default + Eq,
     K: Default + Eq,
     M: HashVec<K, V> + HashVecStatisticsInternal<K, V>,
 {
     hash_vec.get_statistics_mut().size -= 1;
-    let is_max = hash_vec
-        .get_statistics()
-        .is_max_length_bucket(index as usize);
+    let is_max = hash_vec.get_statistics().is_max_length_bucket(index);
     if let Some(true) = is_max {
-        hash_vec.get_statistics_mut().remove_bucket(index as usize);
+        hash_vec.get_statistics_mut().remove_bucket(index);
         if hash_vec.get_statistics().get_count() == 0 {
             let bucket_len = hash_vec.get_bucket_len(index);
             if let Some(len) = bucket_len {
                 hash_vec.get_statistics_mut().update(len - 1);
-                hash_vec.get_statistics_mut().add_bucket(index as usize);
+                hash_vec.get_statistics_mut().add_bucket(index);
             }
         }
     }
